ユーザーページの提出一覧テーブル部分を SubmissionsList に切り出す
提出一覧はユーザーページ以外でも使うことが想定されるため
